backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
deep-translator==1.11.4
pyjwt==2.10.1
bcrypt==4.1.1
